# back/openapi.yaml
openapi: 3.1.0
info:
  title: VLR Backend
  version: 0.1.0
paths:
...
  /api/v1/auth/:
    post:
      tags:
        - Аутентификация
      summary: Authenticate User
      operationId: authenticate_user_api_v1_auth__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRequest"
        required: true
      responses:
        200:
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        422:
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /api/v1/auth/me:
    patch:
      tags:
        - Аутентификация
      summary: Обновить профиль
      operationId: update_profile_api_v1_auth_me_patch
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
        required: true
      responses:
        200:
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        422:
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        - HTTPBearer: []
  /api/v1/auth/dev-login:
    post:
      tags:
        - Аутентификация
      summary: Логин для разработчика (Без проверки подписи)
      operationId: dev_authenticate_api_v1_auth_dev_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DevAuthRequest"
        required: true
      responses:
        200:
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        422:
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /api/v1/attempts/:
    get:
      tags:
        - Попытки
      summary: Список попыток
      operationId: list_attempts_api_v1_attempts__get
      responses:
        200:
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/AttemptListItem"
                type: array
                title: Response List Attempts Api V1 Attempts  Get
      security:
        - HTTPBearer: []
    post:
      tags:
        - Попытки
      summary: Начать выполнение лабы
      operationId: create_attempt_api_v1_attempts__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttemptCreate"
        required: true
      responses:
        200:
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttemptResponse"
        422:
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        - HTTPBearer: []
  /api/v1/attempts/{attempt_id}/finish:
    post:
      tags:
        - Попытки
      summary: Завершить попытку
      operationId: finish_attempt_api_v1_attempts__attempt_id__finish_post
      security:
        - HTTPBearer: []
      parameters:
        - name: attempt_id
          in: path
          required: true
          schema:
            type: integer
            title: Attempt Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttemptFinishRequest"
      responses:
        200:
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttemptResponse"
        422:
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /api/v1/attempts/{attempt_id}/review:
    post:
      tags:
        - Попытки
      summary: Проверить работу (учитель)
      operationId: review_attempt_api_v1_attempts__attempt_id__review_post
      security:
        - HTTPBearer: []
      parameters:
        - name: attempt_id
          in: path
          required: true
          schema:
            type: integer
            title: Attempt Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttemptReviewRequest"
      responses:
        200:
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttemptResponse"
        422:
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /api/v1/attempts/{attempt_id}:
    delete:
      tags:
        - Попытки
      summary: Удалить попытку
      operationId: delete_attempt_api_v1_attempts__attempt_id__delete
      security:
        - HTTPBearer: []
      parameters:
        - name: attempt_id
          in: path
          required: true
          schema:
            type: integer
            title: Attempt Id
      responses:
        204:
          description: Successful Response
        422:
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
...
  /api/v1/classrooms/:
    get:
      tags:
        - Классы
      summary: Классы учителя
      operationId: list_classrooms_api_v1_classrooms__get
      responses:
        200:
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/ClassroomResponse"
                type: array
                title: Response List Classrooms Api V1 Classrooms  Get
      security:
        - HTTPBearer: []
    post:
      tags:
        - Классы
      summary: Создать класс
      operationId: create_classroom_api_v1_classrooms__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassroomCreate"
        required: true
      responses:
        201:
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassroomResponse"
        422:
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        - HTTPBearer: []
  /api/v1/classrooms/advance-year:
    post:
      tags:
        - Классы
      summary: Перевести всех на следующий год
      operationId: advance_year_api_v1_classrooms_advance_year_post
      responses:
        200:
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/ClassroomResponse"
                type: array
                title: Response Advance Year Api V1 Classrooms Advance Year Post
      security:
        - HTTPBearer: []
...
components:
  schemas:
...
    AttemptCreate:
      properties:
        lab_id:
          type: integer
          title: Lab Id
        mode:
          type: string
          enum:
            - test
            - work
          title: Mode
      type: object
      required:
        - lab_id
        - mode
      title: AttemptCreate
    AttemptFinishRequest:
      properties:
        grade:
          anyOf:
            - type: integer
            - type: null
          title: Grade
        result_data:
          anyOf:
            - $ref: "#/components/schemas/AttemptResultData"
            - type: null
        comment:
          anyOf:
            - type: string
            - type: null
          title: Comment
      type: object
      title: AttemptFinishRequest
    AttemptListItem:
      properties:
        id:
          type: integer
          title: Id
        student_id:
          type: integer
          title: Student Id
        lab_id:
          type: integer
          title: Lab Id
        mode:
          type: string
          title: Mode
        status:
          type: string
          title: Status
        result_data:
          anyOf:
            - $ref: "#/components/schemas/AttemptResultData"
            - additionalProperties: true
              type: object
            - type: null
          title: Result Data
        grade:
          anyOf:
            - type: integer
            - type: null
          title: Grade
        comment:
          anyOf:
            - type: string
            - type: null
          title: Comment
        started_at:
          type: string
          format: "date-time"
          title: Started At
        finished_at:
          anyOf:
            - type: string
              format: "date-time"
            - type: null
          title: Finished At
        lab_title:
          type: string
          title: Lab Title
        student_name:
          type: string
          title: Student Name
        class_number:
          anyOf:
            - type: integer
            - type: null
          title: Class Number
      type: object
      required:
        - id
        - student_id
        - lab_id
        - mode
        - status
        - started_at
        - lab_title
        - student_name
      title: AttemptListItem
    AttemptResponse:
      properties:
        id:
          type: integer
          title: Id
        student_id:
          type: integer
          title: Student Id
        lab_id:
          type: integer
          title: Lab Id
        mode:
          type: string
          title: Mode
        status:
          type: string
          title: Status
        result_data:
          anyOf:
            - $ref: "#/components/schemas/AttemptResultData"
            - additionalProperties: true
              type: object
            - type: null
          title: Result Data
        grade:
          anyOf:
            - type: integer
            - type: null
          title: Grade
        comment:
          anyOf:
            - type: string
            - type: null
          title: Comment
        started_at:
          type: string
          format: "date-time"
          title: Started At
        finished_at:
          anyOf:
            - type: string
              format: "date-time"
            - type: null
          title: Finished At
      type: object
      required:
        - id
        - student_id
        - lab_id
        - mode
        - status
        - started_at
      title: AttemptResponse
    AttemptResultData:
      properties:
        measurements:
          items:
            $ref: "#/components/schemas/Measurement"
          type: array
          title: Measurements
        answers:
          additionalProperties: true
          type: object
          title: Answers
        client_metrics:
          additionalProperties: true
          type: object
          title: Client Metrics
      type: object
      title: AttemptResultData
    AttemptReviewRequest:
      properties:
        grade:
          anyOf:
            - type: integer
            - type: null
          title: Grade
        comment:
          anyOf:
            - type: string
            - type: null
          title: Comment
      type: object
      title: AttemptReviewRequest
    AuthRequest:
      properties:
        initData:
          type: string
          title: Initdata
        role:
          type: string
          enum:
            - student
            - teacher
          title: Role
          default: student
      type: object
      required:
        - initData
      title: AuthRequest
    ClassroomCreate:
      properties:
        school:
          anyOf:
            - type: string
            - type: null
          title: School
        grade:
          type: integer
          title: Grade
      type: object
      required:
        - grade
      title: ClassroomCreate
    ClassroomResponse:
      properties:
        id:
          type: integer
          title: Id
        school:
          type: string
          title: School
        grade:
          type: integer
          title: Grade
        label:
          type: string
          title: Label
        students_count:
          type: integer
          title: Students Count
          default: 0
      type: object
      required:
        - id
        - school
        - grade
        - label
      title: ClassroomResponse
...
    DevAuthRequest:
      properties:
        max_user_id:
          type: integer
          title: Max User Id
          default: 123456789
        role:
          type: string
          title: Role
          default: teacher
      type: object
      title: DevAuthRequest
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: "#/components/schemas/ValidationError"
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    Measurement:
      properties:
        step:
          type: integer
          title: Step
        inputs:
          additionalProperties:
            type: number
          type: object
          title: Inputs
        outputs:
          additionalProperties:
            type: number
          type: object
          title: Outputs
      type: object
      required:
        - step
        - inputs
        - outputs
      title: Measurement
    ProfileUpdate:
      properties:
        full_name:
          anyOf:
            - type: string
            - type: null
          title: Full Name
        city:
          anyOf:
            - type: string
            - type: null
          title: City
        school:
          anyOf:
            - type: string
            - type: null
          title: School
        class_number:
          anyOf:
            - type: integer
            - type: null
          title: Class Number
        subject:
          anyOf:
            - type: string
            - type: null
          title: Subject
      type: object
      title: ProfileUpdate
    TokenResponse:
      properties:
        access_token:
          type: string
          title: Access Token
        token_type:
          type: string
          title: Token Type
          default: bearer
        is_new_user:
          type: boolean
          title: Is New User
        user:
          $ref: "#/components/schemas/UserResponse"
      type: object
      required:
        - access_token
        - is_new_user
        - user
      title: TokenResponse
    UserResponse:
      properties:
        max_user_id:
          type: integer
          title: Max User Id
        full_name:
          type: string
          title: Full Name
        role:
          type: string
          enum:
            - student
            - teacher
          title: Role
        city:
          anyOf:
            - type: string
            - type: null
          title: City
        school:
          anyOf:
            - type: string
            - type: null
          title: School
        class_number:
          anyOf:
            - type: integer
            - type: null
          title: Class Number
        subject:
          anyOf:
            - type: string
            - type: null
          title: Subject
        id:
          type: integer
          title: Id
        created_at:
          type: string
          format: "date-time"
          title: Created At
      type: object
      required:
        - max_user_id
        - full_name
        - role
        - id
        - created_at
      title: UserResponse
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer